+++ python/orca/src/bigdl/orca/learn/pytorch/experimential/utils.py
@@ -1,0 +1,43 @@
+#
+# Copyright 2016 The BigDL Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import ray
+from ray.exceptions import RayActorError
+from bigdl.dllib.utils.log4Error import *
+
+def get_driver_node_ip():
+    """
+    Returns the IP address of the current node.
+    :return: the IP address of the current node.
+    """
+    return ray._private.services.get_node_ip_address()
+
+def check_for_failure(remote_values):
+    """Checks remote values for any that returned and failed.
+    :param remote_values: List of object IDs representing functions
+            that may fail in the middle of execution. For example, running
+            a SGD training loop in multiple parallel actor calls.
+    :return Bool for success in executing given remote tasks.
+    """
+    unfinished = remote_values
+    try:
+        while len(unfinished) > 0:
+            finished, unfinished = ray.wait(unfinished)
+            finished = ray.get(finished)
+        return True
+    except RayActorError as exc:
+        logger.exception(str(exc))
+    return False
